fixed import errors caused by changes after merge
