+++ database_tools/tools/__init__.py
@@ -1,3 +1,2 @@
 from .DataLocator import MimicDataLocator
 from .BuildDatabase import BuildDatabase
-from .DataEvaluator import DataEvaluator
